api/requirements.txt
alembic==1.7.3
apturl==0.5.2
arrow==1.2.3
asgiref==3.6.0
attrs==19.3.0
autopep8==1.4.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
binaryornot==0.4.4
blinker==1.4
boto3==1.26.41
botocore==1.29.41
Brlapi==0.7.0
cached-property==1.5.1
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cliapp==1.20180812.1
click==7.1.2
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cookiecutter==2.1.1
cryptography==37.0.2
cssutils==1.0.2
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.4.0
Django==4.1.5
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-storages==1.13.2
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.8.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.4.4
gajim==1.1.3
grpcio==1.16.1
gunicorn==20.0.4
httplib2==0.14.0
idna==3.3
ifaddr==0.1.6
IMDbPY==6.8
importlib-resources==5.2.2
importlib_metadata==7.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==1.0.1
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==4.2.0
nbxmpp==0.6.10
nemo-emblems==4.8.1
netaddr==0.7.19
netifaces==0.10.4
nose==1.3.7
numpy==1.17.4
oauthlib==3.2.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pep8==1.7.1
pexpect==4.6.0
Pillow==7.0.0
pipenv==2023.2.4
platformdirs==2.5.2
precis-i18n==1.0.0
protobuf==3.6.1
psutil==5.5.1
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyparted==3.11.2
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.5
python-apt==2.0.1+ubuntu0.20.4.1
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.15.0
python-gnupg==0.4.5
python-slugify==8.0.1
python-xapp==2.0.2
python-xlib==0.23
python3-openid==3.2.0
pytz==2018.5
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
reportlab==3.5.34
requests==2.28.0
requests-file==1.4.3
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
s3transfer==0.6.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.15.0
soupsieve==1.9.5
SQLAlchemy==1.3.22
sqlparse==0.2.4
ssh-import-id==5.10
systemd-python==234
termcolor==2.3.0
text-unidecode==1.3
texttable==1.6.2
tinycss2==1.0.2
tldextract==2.2.1
ttystatus==0.38
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.26.9
uWSGI==2.0.24
virtualenv==20.16.5
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==1.0.1
WoeUSB-ng @ file:///home/jamie/Documents/MobileDev/WoeUSB-ng
wxPython==4.0.7
xkit==0.0.0
youtube-dl==2021.4.26
zeroconf==0.24.4
zipp==3.5.0
